fix potential unsigened integer overflow
should not have a bad impact on all current platforms, as the
result is properly casted. It's still a potential problem spot.

Detecec by clang UBSAN.

+++ runtime/stringbuf.c
@@ -703,7 +703,7 @@ int rsCStrSzStrCmp(cstr_t *pCS1, uchar *psz, size_t iLenSz)
 		else
 			return strncmp((char*)pCS1->pBuf, (char*)psz, iLenSz);
 	else
-		return pCS1->iStrLen - iLenSz;
+		return (ssize_t) pCS1->iStrLen - (ssize_t) iLenSz;
 }
